Use pimega library from system installation
Expect the library to be already installed in the system directories to
simplify IOC build system and make the library version less coupled with
support module. Since develop cycles happen in a high frequency now,
this is being preferred over importing headers and binaries into a
support module, as done in other areaDetector modules with proprietary
SDKs.

We need to explicitly define include directory for pimega library due to
the usage of `#include ""` directives. By default, we expect headers to
be installed under `pimega` directory, to avoid their generic
subdirectories from conflicting with other application headers. This
should be redefined in CONFIG_SITE.local to override the default value
if needed.