Use standard containerization scheme.
`BUILD_IOCS` had to be configured after Makefile includes, so that the
Area Detector configuration in epics-in-docker is overwritten.

Image name and label metadata have been defined using environment
variables, so that they can be defined using the CI dynamic variables.

ADPimega RELEASE file is now reused in the IOC build, so that the
definitions from epics-in-docker are forwarded to it. This is required
to build the IOC inside the support module in the current configuration.

We are using epics-in-docker v0.5.0 to use NDSSCPimega plugin, which is
linked to the IOC.