Add MagicStart: compatible with GUI refactor.
Some latest GUI refactor take control logic away from GUI.
These changes only make snese in the GUI if the two records
added by this commit exist.

Ideally, these records arent needed for operation, but historically
users have counted on this logic to operate the detector.

The fact that these records are added does not remove EPICS compliance
because it does not change Capture and Acquire records functionality. But
it adds the possibility of orchestrating them to produce a coherent result.