Remove redundant parameters creation
These parameters are already created on `createParameters` in the
pimegaDetector constructor, and do not need to be recreated each time
a request is processed.

Note that these parameters reimplement the asynRecord trace control
fields [1]. A better approach in the long run is actually to remove them
all, and use getTraceMask method from the asynTrace singleton instead.

[1]: https://epics-modules.github.io/asyn/asynRecord.html#trace-control-fields