Wait for detector and backend before changing Acquire_RBV to 1
Writing one (1) to Acquire PV makes a message to sent from the driver
thread, from `pimegaDetector::writeInt32`, to the pimegaDetTask (which
runs `pimegaDetector::acqTask()`) and right away changes the Acquire_RBV
value correspondingly through `setIntegerParam()`. This renders the
Acquire_RBV PV useless for synchronization purposes in higher layers, as
it is not guaranteed that neither the detector is ready for triggers nor
RDMA buffers are properly set by the pimegaDetTask and
pimegacaptureTask, respectively, as no synchronization barrier is
implemented.

Make Acquire_RBV be updated by the pimegaDetTask instead only after the
backend has moved its status to acquiring, which means it is currently
waiting for frames to arrive through RDMA, and chips and gateway have
been correctly configured. Waiting for the backend must occur first, as
it must have buffers properly set before frames are sent through the
network links.

To get the status from the backend, `get_acqStatus_from_backend()` is
used directly, since `status_acquire()` does more than simply reading
the detector state, which is not desired here. The time to wait during
polling has been arbitrarily chosen not to overload the socket and to
prevent from delaying without need.

Since the thread becomes blocked waiting for the status to change,
`startAcquire()` cannot be called from the thread which sends the
acquisition parameters to the backend. This is not a problem here, since
pimegaDetTask is only responsible for configuring the detector, not the
backend.

"Acquiring" string is now only published through the
IOCStatusMessage_RBV PV if the acquisition could be successfully
started. In case of failing to start, it will go directly to the failed
state as already defined in current implementation.