Wait detector and backend stop before changing Acquire_RBV to 0
Stopping acquisition renders Acquire_RBV to go back to zero (0) right
after processing the request, even though messages to the detector and
backend might have not yet sent by `acqTask()`. This causes issues when
relying on readback values to trigger another acquisition in a sequel.

Make Acquire_RBV readback wait for the detector be stopped (through
`stop_acquire()`) and backend receive and reach the done state. This
ensures both that other commands can be sent to the detector afterwards
(without stalling the detector) and that another acquisition can be
configured in the backend (with frames being correctly discarded).

Note that `acqTask()` becomes responsible for both sending stop command
to the detector, which was previously done by `captureTask()`, and also
stopping RDMA acquisition in the backend, but not aborting saving frames
already in the backend buffer. Aborting should still be responsibility
of the pimegacaptureTask, since it is exclusively related to saving
frames to a file.

Return value from `stop_acquire()` is not checked, since it does not
return PIMEGA_SUCCESS upon successful execution, but rather the constant
4.