Wait for backend to be acquiring before changing Capture_RBV to 1
Driver thread will send the message to the backend with acquisition
arguments, but backend might not be actually ready for receiving frames
yet. Make `startCaptureBackend()` wait for the backend status go to
"acquiring" before continuing and setting up the Capture_RBV value
through `setIntegerParam()`.

The current implementation assumes that the backend will change to an
error state which can be captured by the IOC if it will never reach
"acquiring" state. Therefore, this blocks the driver thread for as long
it takes to start up everything in the backend-side.

Note that an ADAcquire 1 will only be successfully processed after
NDFileCapture is one. Therefore, `writeInt32()` must have finished
executing `startCaptureBackend()`, and thus the backend must be in
acquiring state. This implies that it is not possible to acqTask and
captureTask skip seing the "Acquiring" state due to an eventual quick
transition from "done" to "acquiring" and once again back in "done".

Control messages are kept being sent in the driver thread, differently
from the other change requests to ADAcquire and NDCapture, to keep the
diff shorter. It should be considered moving it to captureTask to keep
the pattern and let the waiting happen in another thread.