Wait for backend to stop before changing Capture_RBV to 0
Capture_RBV must go to zero only after abort has sucessfully completed
in the backend side. Change it back to zero in the captureTask, so that
the backend status can be checked after sending the abort command, and
updating the readback value correspondingly.

Existing attempt to synchronize based on SAVEDFRAMENUM going back to
zero does not make sense in current versions of the backend, since it
only resets it to zero when starting another acquisition. Thus, the
status itself is checked instead.

TimeRemaining has been kept in `writeInt32()` to keep the semantics of
the reset value during abort. Ideally, it should be set to zero along
with Acquire_RBV, since its timer corresponds to the time remaining to
the detector read-out the data, not the backend to actually grab them,
and write them to disk. Therefore, it should also be set to zero when an
acquisition is simply stopped, but not aborted.

IOC and "server" status messages have been kept to prevent user
experience regression, even though they do not correspond to the actual
behaviour of the code. Note that other parts of the code, including
`startBackendCapture()` and `captureTask()`, should be changed to make
messages consistent with the actual behaviour.