Use libpimega with error handling in get_acqStatus_from_backend
This version of libpimega better handles errors in
`get_acqStatus_from_backend()`, which is extensively used with the new
synchronization mechanism for correct readback implementation.

Multipart message is removed from the pimega-api to ensure disconnection
between a multipart message does not cause the receiving side to crash.
Remove the topic specification to work with the new API, since single
messages do not have them.

Co-authored-by: Marco A. B. Montevechi Filho <marco.filho@lnls.br>