Reset backend state on permission error
When a permission denied occurs, backend changes its state accordingly.
However, we need to first leave this state before attempting to acquire
once again in order to avoid a race condition, and to properly
discriminate another permission denied from a yet in progress state
transition.

Although we can only have a permission denied error leading to a
asynError, we parse the status explicitly to avoid mistakenly forgetting
this error handling inconditional in the future.