Use logFileName static buffer directly
Conditionally allocating memory for log_file_path leads to a bug when
logging is disabled because pimegaDetector::setDefaults() will later
attempt to set the PimegaLogFile parameter based on this uninitialized
member. Considering that this was introduced by the fixed commit which
aims at hiding the struct fields, and that such hinding no longer makes
sense, simply revert to the original approach of using logFileName
directly.

Fixes: 0d94f9e (Removed direct pimega pointer calls, 2024-02-23)