Fix stack corruption in dacCountScan.
While the dac enum is likely to be already the size and alignment of an
int, that was definitely not the case for current_module, since it is an
uint8_t. That meant the call to getParameter with the "int *" pointer
cast, a case of pointer misuse, was causing stack memory corruption, and
could also lead to unaligned access.