Move module-specific records to dedicated template
Multiple replicated records exist in pimega.template for different
detector modules, but they only cover up to 4 modules. Stop replicating
these by hand, and make use of the substitution feature EPICS base
provides us. Because we load them automatic and unconditionally, we must
provide a substitution file for all models.

Alternatively to the implemented approach, we could provide
model-specific iocsh templates that do the dbLoadRecords call by
themselves, which should be loaded in each user of pimega-template.
Because such script would contain only a dbLoadRecords(), we follow a
simpler path of letting this be replicated. This might be changed later
if needed.

These changes also fix two existing bugs with Temperature_Status PV:
modules 3 and 4 wrong record type (ai instead of mbbi), and the usage of
the non-existent PREC field for mbbi.

[1] https://github.com/epics-base/epics-base/blob/72936844aeb0a58dd2d42c57292f2a64bad88d64/modules/libcom/src/macLib/macCore.c#L734