Document template settings based on environment variables
There are multiple environment variables that must be set in order to
use this template. In order to avoid having users search the entire file
looking for their usage and meaning, move all pimegaDetector parameters
to the top of the file, and make it more focused on the environment
variables to be defined.

In order to keep the information of which parameter it forwards to
pimegaDetectorConfig(), define a syntax to specify this in the list of
parameters.

Notice that variables already defined in envPaths are not documented,
since they should always be loaded and not defined by the users.