Make calibration file path a template variable
These calibration files often differ by name and location in the
filesystem for each deploy. As such, it doesn't make sense to include it
hard-coded in the template file. Make it a proper environment variable
to be defined by the iocsh script that uses the template.

Calibration path defined in each model's main script is merely
illustrative, and should be changed by users in their own instance of
the main iocsh script. To give an educated example, use a FHS compliant
path.