Use better prefixes for each model example
These examples have been replicated based on the 135D model, but their
prefix were not even updated accordingly. Fix that.

While here, also remove the duplicate reference to "PIMEGA" in this
prefix. Instead, use "DET" which can refer both to "detector" in a
general term and, for LNLS, "Detectors Group".