Use asyn parameter addresses for module-specific PVs
Having different asyn parameters for each module makes the code harder
to implement, leading to 4 times more parameters, intermediary mappings,
arithmetics in parameter indices, or even bare code replication. Avoid
all of that by using asyn's address lists to have another dimension for
each parameter. This makes it easier to implement a generic handling for
all modules, and also opens the door to using other EPICS substitutions
to instantiate similar records at build time or runtime. Deduplicating
records is left to its own patch.

Note that, even though some records are currently instantiated up to
four modules, we always define 10 addresses so that we can rely on
max_num_modules value for any detector model. This will make it easier
to later properly export them all to PVs according to the detector model
in use. We must define it statically because the asynPortDriver will
create the parameter lists based on this value in its own constructor
when we still haven't created the base pimega struct containing the
number of modules.

Notice that everything in the driver already relies on the fact that the
maximum value for max_num_modules is 10. Relaxing this assumption is out
of the scope of this patch.

We only call callbacks in all addresses when the code block has updated
any parameter that is module-specific. The exceptions are
TemperatureStatus and TemperatureHighest, which don't propagate read
values by their own.