Loop through all modules using zero-based index
In different places, we loop through the modules from 1 up to the
pimega->max_num_modules. However, what we really need are indices from 0
up to `max_num_modules-1`, as it is evident by the widespread usage of
`module-1` everywhere. Make things simpler and loop directly in the
indices we are interested in.