Update description for plugin iocsh template
This plugin set was imported back in 09e378f (Added commons plugins, and
corrected corruption in array data. was deleting the buffer that was
continously being used when the reference count reaches 0, 2020-12-15),
and its description was never updated according to its usage here. We
won't use it as an example, but instead as the actual set of plugins to
be used at runtime, since we load it in the pimega-template.cmd.

Also, we are not loading autosave requirements here. Remove those
references from the description.