Define max image dimensions close to detector model
Although listed under the "Plugins" section, this variable is also used
by the pimegaDetector command to instantiate the detector driver. Move
it to a more general place to avoid misguiding readers. Having this
close to the detector model is interesting because maximum image size
will generally be in function of the model.