Define model based on its name in iocsh scripts
Make the declaration of the model more intuitive by making it based on
the model name instead of the internal pimega library enum mapping. To
do so, we use an indirect evaluation of the model number. This way, we
only need to keep this mapping in the template.

In case the detector model name has a typo, or doesn't exist at all, the
expansion in pimegaDetectorConfig() will fail, leading to a shell halt.
Thus, this issue should be easy to spot and fix during the first tests
in a new deploy.