Halt on any non-interactive iocsh error
When loading the startup script, there might be typos in the settings
that leads the IOC into an inoperable state. This will be even more
relevant for upcoming changes.

Make iocsh halt when errors occur in order to make this situation easier
to detect and debug. This way, when starting the IOC, the last message
displayed will be the error that occurred during the initialization.
Because it will halt instead of crashing, process managers (such as
procServ) won't attempt to continuously restart it regardless of its
specific settings.

This option doesn't apply to interactive sessions, so there is no issue
with mistyping commands in the shell after the IOC is up and running.