Don't document Mobipix as a supported detector
While libpimega defines Mobipix as an option in its
pimega_detector_model_t enum, it actually doesn't support this detector,
which has its own dedicated IOC. Don't document this as a valid option
anymore, so that users have an assertive list of supported models. This
also removes the need to create more boilerplate in future changes for a
model that can't be used.